chore: rename file to autocompleteConstants and move it to codemirror utils
