+++ packages/bruno-app/src/components/CollectionSettings/Headers/index.js
@@ -13,7 +13,7 @@ import { saveCollectionRoot } from 'providers/ReduxStore/slices/collections/acti
 import SingleLineEditor from 'components/SingleLineEditor';
 import StyledWrapper from './StyledWrapper';
 import { headers as StandardHTTPHeaders } from 'know-your-http-well';
-import { MimeTypes } from 'utils/collections/mimeTypes';
+import { MimeTypes } from 'utils/codemirror/autocompleteConstants';
 const headerAutoCompleteList = StandardHTTPHeaders.map((e) => e.header);
 
 const Headers = ({ collection }) => {
