+++ packages/bruno-app/src/utils/codemirror/autocompleteConstants.js
